# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/categories/{id}": {
      "get": {
        "operationId": "CategoryController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "tags": ["categories"],
        "security": [{ "bearer": [] }]
      },
      "put": {
        "operationId": "CategoryController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateCategoryDto" }
            }
          }
        },
        "responses": { "200": { "description": "" } },
        "tags": ["categories"],
        "security": [{ "bearer": [] }]
      },
      "delete": {
        "operationId": "CategoryController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "tags": ["categories"],
        "security": [{ "bearer": [] }]
      }
    },
    "/subcategories": {
      "get": {
        "operationId": "SubcategoryController_findAll",
        "parameters": [],
        "responses": { "200": { "description": "" } },
        "tags": ["subcategories"]
      },
      "post": {
        "operationId": "SubcategoryController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateSubcategoryDto" }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "tags": ["subcategories"],
        "security": [{ "bearer": [] }]
      }
    },
    "/subcategories/{id}": {
      "get": {
        "operationId": "SubcategoryController_findById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "tags": ["subcategories"]
      },
      "put": {
        "operationId": "SubcategoryController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateSubcategoryDto" }
            }
          }
        },
        "responses": { "200": { "description": "" } },
        "tags": ["subcategories"],
        "security": [{ "bearer": [] }]
      },
      "delete": {
        "operationId": "SubcategoryController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "tags": ["subcategories"],
        "security": [{ "bearer": [] }]
      }
    },
...
    "/reduction-codes": {
      "post": {
        "operationId": "ReductionCodeController_createReductionCode",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReductionCodeDto"
              }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "tags": ["reduction-codes"],
        "security": [{ "bearer": [] }]
      },
      "get": {
        "operationId": "ReductionCodeController_getAllReductionCodes",
        "parameters": [],
        "responses": { "200": { "description": "" } },
        "tags": ["reduction-codes"],
        "security": [{ "bearer": [] }]
      }
    },
    "/reduction-codes/{code}": {
      "get": {
        "operationId": "ReductionCodeController_getReductionCodeByCode",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "tags": ["reduction-codes"]
      },
      "put": {
        "operationId": "ReductionCodeController_updateReductionCode",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReductionCodeDto"
              }
            }
          }
        },
        "responses": { "200": { "description": "" } },
        "tags": ["reduction-codes"],
        "security": [{ "bearer": [] }]
      },
      "delete": {
        "operationId": "ReductionCodeController_deleteReductionCode",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "tags": ["reduction-codes"],
        "security": [{ "bearer": [] }]
      }
    },
...
  },
  "info": {
    "title": "API",
    "description": "",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": { "scheme": "bearer", "bearerFormat": "JWT", "type": "http" }
    },
    "schemas": {
...
      "UpdateCategoryDto": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "image": { "type": "string" },
          "subcategories": {
            "description": "Array of Subcategory IDs",
            "type": "array",
            "items": { "type": "string" }
          }
        },
        "required": ["name", "image", "subcategories"]
      },
      "CreateSubcategoryDto": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "categoryId": { "type": "string", "description": "Category ID" }
        },
        "required": ["name", "categoryId"]
      },
      "UpdateSubcategoryDto": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "categoryId": { "type": "string", "description": "Category ID" }
        },
        "required": ["name", "categoryId"]
      },
...
      "CreateReductionCodeDto": { "type": "object", "properties": {} },
      "UpdateReductionCodeDto": { "type": "object", "properties": {} },
...
    }
  }
}